+++ R/cifreg.R
@@ -584,4 +584,369 @@ res <- .Call("S0_FG_GcR",as.double(S0),as.double(Gct),
 return(res)
 }# }}}
 
+##' @export
+strataC <- survival:::strata
+
+##' Augmentation for Fine-Gray model based on stratified NPMLE Cif (Aalen-Johansen) 
+##'
+##' Computes  the augmentation term for each individual as well as the sum
+##' \deqn{
+##' A(\beta) = \int H(t,X,\beta) \frac{F_2^*(t,s)}{S^*(t,s)} \frac{1}{G_c(t)} dM_c
+##' }
+##' with 
+##' \deqn{
+##' H(t,X,\beta) = \int_t^\infty (X - E(\beta,t) ) G_c(t) d\Lambda_1^*i(t,s)
+##' }
+##' using a KM for \deqn{G_c(t)} and a working model for cumulative baseline
+##' related to \deqn{F_1^*(t,s)} and \deqn{s} is strata, 
+##' \deqn{S^*(t,s) = 1 - F_1^*(t,s) - F_2^*(t,s)}, and
+##' \deqn{E(\beta^p,t)} is given.
+##'
+##' After a couple of iterations we end up with a solution of 
+##' \deqn{
+##' \int (X - E(\beta) ) Y_1(t) w(t) dM_1 + A(\beta)
+##' }
+##' the augmented FG-score. 
+##'
+##' Standard errors computed under assumption of correct \deqn{G_c} model.
+##'
+##' @param formula formula with 'Event', strata model for CIF given by strata, and strataC specifies censoring strata
+##' @param data data frame
+##' @param offset offsets for cox model
+##' @param data data frame
+##' @param E from FG-model 
+##' @param cause of interest 
+##' @param cens.code code of censoring 
+##' @param km to use Kaplan-Meier
+##' @param case.weights weights for FG score equations (that follow dN_1) 
+##' @param weights weights for FG score equations
+##' @param offset offsets for FG   model
+##' @param ... Additional arguments to lower level funtions
+##' @author Thomas Scheike
+##' @examples
+##' rho1 <- 0.2; rho2 <- 0.9
+##' n <- 200
+##' beta=c(0.0,-0.1,-0.5,0.3)
+##' dats <- simul.cifs(n,rho1,rho2,beta)
+##' dsort(dats) <- ~time
+##' fg <- cifreg(Event(time,status)~Z1+Z2,data=dats,cause=1,propodds=NULL)
+##' 
+##' fgaugS <- FG_AugmentCifstrata(Event(time,status)~Z1+Z2+strata(Z1,Z2),data=dats,cause=1,E=fg$E)
+##' fgaugS2 <- FG_AugmentCifstrata(Event(time,status)~Z1+Z2+strata(Z1,Z2),data=dats,cause=1,E=fgaugS$E)
+##' 
+##' @aliases strataC  simul.cifs
+##' @export
+FG_AugmentCifstrata <- function(formula,data=data,E=NULL,cause=NULL,cens.code=0,km=TRUE,case.weights=NULL,weights=NULL,offset=NULL,...)
+{# {{{
+  cl <- match.call()
+  m <- match.call(expand.dots = TRUE)[1:3]
+  special <- c("strata", "cluster","offset","strataC")
+  Terms <- terms(formula, special, data = data)
+  m$formula <- Terms
+  m[[1]] <- as.name("model.frame")
+  m <- eval(m, parent.frame())
+  Y <- model.extract(m, "response")
+  if (class(Y)!="Event") stop("Expected a 'Event'-object")
+  if (ncol(Y)==2) {
+    exit <- Y[,1]
+    entry <- NULL ## rep(0,nrow(Y))
+    status <- Y[,2]
+  } else {
+    entry <- Y[,1]
+    exit <- Y[,2]
+    status <- Y[,3]
+  }
+  id <- strata <- NULL
+  if (!is.null(attributes(Terms)$specials$cluster)) {
+    ts <- survival::untangle.specials(Terms, "cluster")
+    Terms  <- Terms[-ts$terms]
+    id <- m[[ts$vars]]
+  }
+  if (!is.null(stratapos <- attributes(Terms)$specials$strata)) {
+    ts <- survival::untangle.specials(Terms, "strata")
+    Terms  <- Terms[-ts$terms]
+    strata <- m[[ts$vars]]
+    strata.name <- ts$vars
+  }  else strata.name <- NULL
+  if (!is.null(stratapos <- attributes(Terms)$specials$strataC)) {
+    ts <- survival::untangle.specials(Terms, "strataC")
+    Terms  <- Terms[-ts$terms]
+    strataC <- as.numeric(m[[ts$vars]])-1
+    strataC.name <- ts$vars
+  }  else { strataC <- NULL; strataC.name <- NULL}
+
+  if (!is.null(offsetpos <- attributes(Terms)$specials$offset)) {
+    ts <- survival::untangle.specials(Terms, "offset")
+    Terms  <- Terms[-ts$terms]
+    offset <- m[[ts$vars]]
+  }  
+  X <- model.matrix(Terms, m)
+  if (!is.null(intpos  <- attributes(Terms)$intercept))
+  X <- X[,-intpos,drop=FALSE]
+  if (ncol(X)==0) X <- matrix(nrow=0,ncol=0)
+  id.orig <- id; 
+  if (!is.null(id)) {
+	  ids <- sort(unique(id))
+	  nid <- length(ids)
+      if (is.numeric(id)) id <-  fast.approx(ids,id)-1 else  {
+      id <- as.integer(factor(id,labels=seq(nid)))-1
+     }
+   } else id <- as.integer(seq_along(exit))-1; 
+
+  p <- ncol(X)
+  beta <- NULL
+  if (is.null(beta)) beta <- rep(0,p)
+  if (p==0) X <- cbind(rep(0,length(exit)))
+
+  if (is.null(strata)) { strata <- rep(0,length(exit)); nstrata <- 1; strata.level <- NULL; } else {
+	  strata.level <- levels(strata)
+	  ustrata <- sort(unique(strata))
+	  nstrata <- length(ustrata)
+	  strata.values <- ustrata
+      if (is.numeric(strata)) strata <-  fast.approx(ustrata,strata)-1 else  {
+      strata <- as.integer(factor(strata,labels=seq(nstrata)))-1
+    }
+  }
+
+if (is.null(strataC)) { strataC <- rep(0,length(exit)); nstrataC <- 1; strataC.level <- NULL; } else {
+	  strataC.level <- levels(strataC)
+	  ustrataC <- sort(unique(strataC))
+	  nstrataC <- length(ustrataC)
+	  strataC.values <- ustrataC
+      if (is.numeric(strataC)) strataC <-  fast.approx(ustrataC,strataC)-1 else  {
+      strataC <- as.integer(factor(strataC,labels=seq(nstrataC)))-1
+    }
+  }
+
+  cens.strata <- strataC
+  cens.nstrata <- nstrataC 
+
+  if (is.null(offset)) offset <- rep(0,length(exit)) 
+  if (is.null(weights)) weights <- rep(1,length(exit)) 
+  strata.call <- strata
+  Z <- NULL
+  Zcall <- matrix(1,1,1) ## to not use for ZX products when Z is not given 
+  if (!is.null(Z)) Zcall <- Z
+
+  ## possible casewights to use for bootstrapping and other things
+  if (is.null(case.weights)) case.weights <- rep(1,length(exit)) 
+
+  trunc <- (!is.null(entry))
+  if (!trunc) entry <- rep(0,length(exit))
+  call.id <- id
+
+  if (!is.null(id)) {
+	  ids <- unique(id)
+	  nid <- length(ids)
+      if (is.numeric(id)) id <-  fast.approx(ids,id)-1 else  {
+      id <- as.integer(factor(id,labels=seq(nid)))-1
+     }
+   } else id <- as.integer(seq_along(entry))-1; 
+   ## orginal id coding into integers 
+   id.orig <- id+1; 
+
+
+  statusC <- (status==cens.code) 
+  statusE <- (status==cause) 
+  if (sum(statusE)==0) stop("No events of type 1 before time \n"); 
+
+  ## sorting after time and statusC, but event times unique after order
+  Zcall <- cbind(status,strata)
+  dd <- .Call("FastCoxPrepStrata",entry,exit,statusC,X,id, 
+	     trunc,strataC,weights,offset,Zcall,case.weights,PACKAGE="mets")
+
+  Z <- dd$X
+  jumps <- dd$jumps+1
+  xxstrataC <- c(dd$strata)
+  xxstatus  <- dd$Z[,1]
+  xxstrata  <- dd$Z[,2]
+  other <- which((!(xxstatus %in% c(cens.code,cause)) ) )               
+  jumps1 <- which(xxstatus==cause)
+  jumpsD <- which(xxstatus!=cens.code)
+  rr <- c(dd$sign*exp(dd$offset))
+  ## S0 after strata
+  S0 = c(revcumsumstrata(rr,strata,nstrata))
+  ## S0 after strataC
+  S00C = c(revcumsumstrata(rr,xxstrataC,nstrataC))
+
+  ## censoring MG, strataC
+  stratJumps <- dd$strata[jumps]
+  S00i <- rep(0,length(dd$strata))
+  S00i[jumps] <-  1/S00C[jumps]
+
+  ## cif calculation, uses strata {{{
+  S0Di <- S02i <- S01i <- rep(0,length(dd$strata))
+  S01i[jumps1] <-  1/S0[jumps1]
+  S02i[other] <-  1/S0[other]
+  S0Di[jumpsD] <-  1/S0[jumpsD]
+
+  ## strata-Cif G_T(t)
+  if (!km) { 
+     cumhazD <- cumsumstratasum(S0Di,xxstrata,nstrata)
+     Stm <- exp(-cumhazD$lagsum)
+     St <- exp(-cumhazD$sum)
+  } else { 
+     StA <- cumsumstratasum(log(1-S0Di),xxstrata,nstrata)
+     Stm <- exp(StA$lagsum)
+     St <- exp(StA$sum)
+  }
+
+  ## G_c(t-) 
+  if (!km) { 
+    cumhazD <- c(cumsumstratasum(S00i,xxstrataC,nstrataC)$lagsum)
+    Gc      <- exp(-cumhazD)
+  } else Gc <- c(exp(cumsumstratasum(log(1-S00i),xxstrataC,nstrataC)$lagsum))
+ cif1 <- cumsumstrata(Stm*S01i,xxstrata,nstrata)
+ cif2 <- cumsumstrata(Stm*S02i,xxstrata,nstrata)
+# }}}
+
+ Et <- matrix(0,nrow(Z),ncol(Z))
+ Et[jumps1,] <- E
+
+ Lam1fg <- -log(1-cif1)
+ laststrata <- tailstrata(xxstrata,nstrata)
+ gtstart <- Lam1fg[laststrata]
+ dLam1fg <- c(diffstrata(Lam1fg,xxstrata,nstrata))
+
+ ## Gc~strataC, dLam1fg~strata
+ tailcstrata <- tailstrata(xxstrataC,nstrataC)
+ Gcstart <- Gc[tailcstrata]
+
+ dstrata <- mystrata(data.frame(cbind(xxstrataC,xxstrata)))
+ ndstrata <- attr(dstrata,"nlevel")
+ lastt <- tailstrata(dstrata-1,ndstrata)
+
+ ### ## \int_t^\infty G_c^j(t) d\Lambda_1^k(t)
+ G0start <- rep(1,nstrataC)
+ cLam1fg  <- cumsum2strata(Gc,dLam1fg,xxstrataC,nstrataC,xxstrata,nstrata,G0start)$res
+ lastt <- tailstrata(dstrata-1,ndstrata)
+ RLam1fg <- cLam1fg[lastt][dstrata]-cLam1fg
+
+ ## E(s) from FG without strata
+ ## \int_0^t  G_c^j(s) E(s) d\Lambda_1^k(s)
+ fff <- function(x) {
+   cx  <- cumsum2strata(Gc,x*dLam1fg,xxstrataC,nstrataC,xxstrata,nstrata,G0start)$res
+    return(cx[lastt][dstrata]-cx)
+ }
+ ERLam1fg0  <- apply(Et,2,fff)
+
+ gt <-  RLam1fg*c(cif2/(Gc*St))
+ gt[gt==Inf] <- 0
+ gt[is.na(gt)] <- 0
+ ERLam1fg  <- ERLam1fg0*c(cif2/(Gc*St))
+ ERLam1fg[ERLam1fg==Inf] <- 0
+ ERLam1fg[is.na(ERLam1fg)] <- 0
+
+ sss <- headstrata(dstrata-1,ndstrata)
+ gtstart <- gt[sss]
+ E1dLam0 <- cumsum2strata(gt,S00i,dstrata-1,ndstrata,xxstrataC,nstrataC,gtstart)$res
+
+ fff <- function(x) {
+    gtstart <- x[sss]
+    cx  <- cumsum2strata(x,S00i,dstrata-1,ndstrata,xxstrataC,nstrataC,gtstart)$res
+    return(cx)
+ }
+ E2dLam0 <- apply(ERLam1fg,2,fff)
+
+ U1 <- matrix(0,nrow(Z),1)
+ U2 <- matrix(0,nrow(Z),ncol(Z))
+ U1[jumps,] <- gt[jumps]
+ U2[jumps,] <- ERLam1fg[jumps,]
+
+ ### Martingale  as a function of time and for all subjects to handle strata 
+ MG1t <- Z*c(U1[,,drop=FALSE]-E1dLam0)*rr*c(dd$weights)
+ MG2t <- (U2[,,drop=FALSE]-E2dLam0)*rr*c(dd$weights)
+ MGt <- MG1t-MG2t
+ MGiid <- apply(MGt,2,sumstrata,dd$id,max(dd$id)+1)
+ augment <- apply(MGt,2,sum)
+
+ augment <- list(MGiid=MGiid,augment=augment,id=id,id.orig=id.orig,
+###  dd=dd,div=div,Et=Et,E=E, gt=gt, ERLam1fg=ERLam1fg,RLam1fg=RLam1fg,ERLam1fg0=ERLam1fg0,
+     jumps1=jumps1,jumps=jumps,other=other,
+     nstrata=nstrata,nstrataC=nstrataC,dstrata=dstrata,ndstrata=ndstrata,
+     cif1=cif1,cif2=cif2,St=St,Gc=Gc,strata=xxstrata,strataC=xxstrataC,time=dd$time)
+
+ ## drop strata's from formula and run wiht augmention term
+ # {{{
+ drop.strata <- function(x) {
+   mm <- unlist(Specials(x,"strata"))
+   for (i in mm) x <- update(x, paste(".~.-strata(",i,")"))
+   mm <- unlist(Specials(x,"strataC"))
+   for (i in mm) x <- update(x, paste(".~.-strataC(",i,")"))
+   return(x)
+ }
+
+ formulans <- drop.strata(formula)
+# }}}
+
+  if (nstrataC==1) cens.model <- ~+1 else cens.model <- ~strata(strataCC)
+  data$strataCC <- cens.strata
+
+ fga <- cifreg(formulans,data=data,cause=cause,
+     propodds=NULL,augmentation=augment$augment,cens.model=cens.model,...)
+
+ ## adjust SE and var based on augmentation term
+ fga$var.orig <- fga$var
+ fga$augment <- augment$augment
+ fga$iid <- fga$iid + MGiid %*% fga$ihessian 
+ fga$var <- crossprod(fga$iid)
+ fga$se.coef <-  diag(fga$var)^.5
+ fga$MGciid <- MGiid
+
+ return(fga)
+}# }}})
+
+##' @export
+simul.cifs <- function(n,rho1,rho2,beta,rc=0.5,depcens=0,rcZ=0.5,bin=1,type=
+		       c("cloglog","logistic"),rate=1) {# {{{
+p=length(beta)/2
+tt <- seq(0,6,by=0.1)
+if (length(rate)==1) rate <- rep(rate,2)
+Lam1 <- rho1*(1-exp(-tt/rate[1]))
+Lam2 <- rho2*(1-exp(-tt/rate[2]))
+
+if (length(bin)==1) bin <- rep(bin,2)
+
+Z=cbind((bin[1]==1)*(2*rbinom(n,1,1/2)-1)+(bin[1]==0)*rnorm(n),(bin[2]==1)*(rbinom(n,1,1/2))+(bin[2]==0)*rnorm(n))
+colnames(Z) <- paste("Z",1:2,sep="")
+
+cif1 <- setup.cif(cbind(tt,Lam1),beta[1:2],Znames=colnames(Z),type=type[1])
+cif2 <- setup.cif(cbind(tt,Lam2),beta[3:4],Znames=colnames(Z),type=type[1])
+###
+data <- sim.cifsRestrict(list(cif1,cif2),n,Z=Z)
+
+if (depcens==0) censor=pmin(rexp(n,1)*(1/rc),6) else censor=pmin(rexp(n,1)*(1/(rc*exp(rcZ*Z[,1]))),6)
+
+status=data$status*(data$time<=censor)
+time=pmin(data$time,censor)
+data <- data.frame(time=time,status=status)
+return(cbind(data,Z))
+
+}# }}}
+
+simul.mod <- function(n,rho1,rho2,beta,rc=0.5,k=1,depcens=0) {# {{{
+p=length(beta)/2
+tt <- seq(0,6,by=0.1)
+Lam1 <- rho1*(1-exp(-tt))
+Lam2 <- rho2*(1-exp(-tt))
+
+Z=cbind(2*rbinom(n,1,1/2)-1,rnorm(n))
+colnames(Z) <- paste("Z",1:2,sep="")
+cif1 <- setup.cif(cbind(tt,Lam1),beta[1:2],Znames=colnames(Z),type="cloglog")
+cif2 <- setup.cif(cbind(tt,Lam2),beta[3:4],Znames=colnames(Z),type="cloglog")
+###
+data <- sim.cifsRestrict(list(cif1,cif2),n,Z=Z)
+
+censhaz  <-  cbind(tt,k*tt)
+if (depcens==1) {
+datc <- rchaz(censhaz,exp(Z[,1]*rc))
+} else datc <- rchaz(censhaz,n=n)
+
+data$time <- pmin(data$time,datc$time)
+data$status <- ifelse(data$time<datc$time,data$status,0)
+dsort(data) <- ~time
+
+return(data)
+}# }}}
+